+++ src/Smithbox.Program/Editors/Map Editor/Data/LightProbeBank.cs
@@ -5,6 +5,7 @@ using StudioCore.Logger;
 using StudioCore.Utilities;
 using System;
 using System.Collections.Generic;
+using System.Diagnostics.Eventing.Reader;
 using static SoulsFormats.NVA;
 
 
@@ -38,6 +39,8 @@ public class LightProbeBank
         if (!CanUse())
             return;
 
+        var fs = View.Project.Handler.MapData.PrimaryBank.TargetFS;
+
         if (View.Project.Descriptor.ProjectType is ProjectType.DS2 or ProjectType.DS2S)
         {
             foreach (var entry in Project.Locator.LightProbeFiles.Entries)
@@ -45,24 +48,31 @@ public class LightProbeBank
                 var bhdPath = entry.Path;
                 var bdtPath = $"{bhdPath}".Replace(".gibhd", ".gibdt");
 
-                try
+                if (fs.FileExists(bdtPath) && fs.FileExists(bhdPath))
                 {
-                    var bdtFile = (Memory<byte>)View.Project.Handler.MapData.PrimaryBank.TargetFS.ReadFile(bdtPath);
-                    var bhdFile = (Memory<byte>)View.Project.Handler.MapData.PrimaryBank.TargetFS.ReadFile(bhdPath);
-
-                    using var bdt = BXF4.Read(bhdFile, bdtFile);
-                    BinderFile file = bdt.Files.Find(f => f.Name.EndsWith("pointcloud.btpb.dcx"));
-
-                    if (file != null)
+                    try
                     {
-                        var btpbData = BTPB.Read(file.Bytes);
+                        var bdtFile = (Memory<byte>)fs.ReadFile(bdtPath);
+                        var bhdFile = (Memory<byte>)fs.ReadFile(bhdPath);
 
-                        Files.Add(entry.Filename, btpbData);
+                        using var bdt = BXF4.Read(bhdFile, bdtFile);
+                        BinderFile file = bdt.Files.Find(f => f.Name.EndsWith("pointcloud.btpb.dcx"));
+
+                        if (file != null)
+                        {
+                            var btpbData = BTPB.Read(file.Bytes);
+
+                            Files.Add(entry.Filename, btpbData);
+                        }
+                    }
+                    catch (Exception e)
+                    {
+                        Smithbox.LogError(this, LOC.Get("MAP_Data_Failed_Read_GI", bhdPath), e);
                     }
                 }
-                catch (Exception e)
+                else
                 {
-                    Smithbox.LogError(this, LOC.Get("MAP_Data_Failed_Read_GI", bhdPath), e);
+                    Smithbox.LogError(this, LOC.Get("MAP_Data_Failed_Find_GI_File", bhdPath));
                 }
             }
         }
@@ -70,20 +80,27 @@ public class LightProbeBank
         {
             foreach (var entry in Project.Locator.LightProbeFiles.Entries)
             {
-                var fileData = Project.VFS.FS.ReadFile(entry.Path);
-
-                if (fileData != null)
+                if (fs.FileExists(entry.Path))
                 {
-                    try
-                    {
-                        var btpbData = BTPB.Read(fileData.Value);
+                    var fileData = fs.ReadFile(entry.Path);
 
-                        Files.Add(entry.Filename, btpbData);
-                    }
-                    catch (Exception e)
+                    if (fileData != null)
                     {
-                        Smithbox.LogError(this, LOC.Get("MAP_Data_Failed_Read_BTPB", entry.Path), e);
+                        try
+                        {
+                            var btpbData = BTPB.Read(fileData.Value);
+
+                            Files.Add(entry.Filename, btpbData);
+                        }
+                        catch (Exception e)
+                        {
+                            Smithbox.LogError(this, LOC.Get("MAP_Data_Failed_Read_BTPB", entry.Path), e);
+                        }
                     }
+                }
+                else
+                {
+                    Smithbox.LogError(this, LOC.Get("MAP_Data_Failed_Find_BTPB", entry.Path));
                 }
             }
         }
@@ -131,6 +148,8 @@ public class LightProbeBank
         if (!CanUse())
             return;
 
+        var fs = View.Project.Handler.MapData.PrimaryBank.TargetFS;
+
         if (View.Project.Descriptor.ProjectType is ProjectType.DS2 or ProjectType.DS2S)
         {
             foreach (var entry in Project.Locator.LightProbeFiles.Entries)
@@ -143,59 +162,73 @@ public class LightProbeBank
                 var bhdPath = entry.Path;
                 var bdtPath = $"{bhdPath}".Replace(".gibhd", ".gibdt");
 
-                try
+                if (fs.FileExists(bdtPath) && fs.FileExists(bhdPath))
                 {
-                    var bdtFile = (Memory<byte>)View.Project.Handler.MapData.PrimaryBank.TargetFS.ReadFile(bdtPath);
-                    var bhdFile = (Memory<byte>)View.Project.Handler.MapData.PrimaryBank.TargetFS.ReadFile(bhdPath);
-
-                    using var bdt = BXF4.Read(bhdFile, bdtFile);
-                    BinderFile file = bdt.Files.Find(f => f.Name.EndsWith("pointcloud.btpb.dcx"));
-
-                    if (file != null)
+                    try
                     {
-                        var applyEdit = false;
+                        var bdtFile = (Memory<byte>)fs.ReadFile(bdtPath);
+                        var bhdFile = (Memory<byte>)fs.ReadFile(bhdPath);
 
-                        foreach (var parent in map.LightProbeParents)
+                        using var bdt = BXF4.Read(bhdFile, bdtFile);
+                        BinderFile file = bdt.Files.Find(f => f.Name.EndsWith("pointcloud.btpb.dcx"));
+
+                        if (file != null)
                         {
-                            // Match the root object name to the filename
-                            if (parent.WrappedObject.ToString() == entry.Filename)
+                            var applyEdit = false;
+
+                            foreach (var parent in map.LightProbeParents)
                             {
-                                var btpbData = BTPB.Read(file.Bytes);
-
-                                // Clear groups and then re-fill from the map container hierarchy
-                                btpbData.Groups.Clear();
-
-                                foreach (var btpbEntry in parent.Children)
+                                // Match the root object name to the filename
+                                if (parent.WrappedObject.ToString() == entry.Filename)
                                 {
-                                    var group = (BTPB.Group)btpbEntry.WrappedObject;
+                                    try
+                                    {
+                                        var btpbData = BTPB.Read(file.Bytes);
 
-                                    btpbData.Groups.Add(group);
-                                }
+                                        // Clear groups and then re-fill from the map container hierarchy
+                                        btpbData.Groups.Clear();
 
-                                var fileOutput = btpbData.Write();
+                                        foreach (var btpbEntry in parent.Children)
+                                        {
+                                            var group = (BTPB.Group)btpbEntry.WrappedObject;
 
-                                if (!BytePerfectHelper.Md5Equal(file.Bytes.Span, fileOutput))
-                                {
-                                    applyEdit = true;
-                                }
+                                            btpbData.Groups.Add(group);
+                                        }
 
-                                if (applyEdit)
-                                {
-                                    file.Bytes = fileOutput;
+                                        var fileOutput = btpbData.Write();
+
+                                        if (!BytePerfectHelper.Md5Equal(file.Bytes.Span, fileOutput))
+                                        {
+                                            applyEdit = true;
+                                        }
+
+                                        if (applyEdit)
+                                        {
+                                            file.Bytes = fileOutput;
+                                        }
+                                    }
+                                    catch (Exception e)
+                                    {
+                                        Smithbox.LogError(this, LOC.Get("MAP_Data_Failed_Read_BTPB", entry.Filename), e);
+                                    }
                                 }
                             }
-                        }
 
-                        if (applyEdit)
-                        {
-                            Project.VFS.ProjectFS.WriteFile(bhdPath, bhdFile.ToArray());
-                            Project.VFS.ProjectFS.WriteFile(bdtPath, bdtFile.ToArray());
+                            if (applyEdit)
+                            {
+                                Project.VFS.ProjectFS.WriteFile(bhdPath, bhdFile.ToArray());
+                                Project.VFS.ProjectFS.WriteFile(bdtPath, bdtFile.ToArray());
+                            }
                         }
                     }
+                    catch (Exception e)
+                    {
+                        Smithbox.LogError(this, LOC.Get("MAP_Data_Failed_Write_GI", bhdPath), e);
+                    }
                 }
-                catch (Exception e)
+                else
                 {
-                    Smithbox.LogError(this, LOC.Get("MAP_Data_Failed_Write_GI", bhdPath), e);
+                    Smithbox.LogError(this, LOC.Get("MAP_Data_Failed_Find_GI_File", bhdPath));
                 }
             }
         }
@@ -207,49 +240,63 @@ public class LightProbeBank
                 if (!entry.Filename.Contains(map.Name))
                     continue;
 
-                var fileData = Project.VFS.FS.ReadFile(entry.Path);
-
-                if (fileData != null)
+                if (fs.FileExists(entry.Path))
                 {
-                    var applyEdit = false;
+                    var fileData = fs.ReadFile(entry.Path);
 
-                    try
+                    if (fileData != null)
                     {
-                        foreach (var parent in map.LightProbeParents)
+                        var applyEdit = false;
+
+                        try
                         {
-                            // Match the root object name to the filename
-                            if (parent.WrappedObject.ToString() == entry.Filename)
+                            foreach (var parent in map.LightProbeParents)
                             {
-                                var btpbData = BTPB.Read(fileData.Value);
-
-                                // Clear groups and then re-fill from the map container hierarchy
-                                btpbData.Groups.Clear();
-
-                                foreach (var btpbEntry in parent.Children)
+                                // Match the root object name to the filename
+                                if (parent.WrappedObject.ToString() == entry.Filename)
                                 {
-                                    var group = (BTPB.Group)btpbEntry.WrappedObject;
+                                    try
+                                    {
+                                        var btpbData = BTPB.Read(fileData.Value);
 
-                                    btpbData.Groups.Add(group);
-                                }
+                                        // Clear groups and then re-fill from the map container hierarchy
+                                        btpbData.Groups.Clear();
 
-                                var fileOutput = btpbData.Write();
+                                        foreach (var btpbEntry in parent.Children)
+                                        {
+                                            var group = (BTPB.Group)btpbEntry.WrappedObject;
 
-                                if (!BytePerfectHelper.Md5Equal(fileData.Value.Span, fileOutput))
-                                {
-                                    applyEdit = true;
-                                }
+                                            btpbData.Groups.Add(group);
+                                        }
 
-                                if (applyEdit)
-                                {
-                                    Project.VFS.ProjectFS.WriteFile(entry.Path, fileOutput);
+                                        var fileOutput = btpbData.Write();
+
+                                        if (!BytePerfectHelper.Md5Equal(fileData.Value.Span, fileOutput))
+                                        {
+                                            applyEdit = true;
+                                        }
+
+                                        if (applyEdit)
+                                        {
+                                            Project.VFS.ProjectFS.WriteFile(entry.Path, fileOutput);
+                                        }
+                                    }
+                                    catch (Exception e)
+                                    {
+                                        Smithbox.LogError(this, LOC.Get("MAP_Data_Failed_Read_BTPB", entry.Path), e);
+                                    }
                                 }
                             }
                         }
+                        catch (Exception e)
+                        {
+                            Smithbox.LogError(this, LOC.Get("MAP_Data_Failed_Write_BTPB", entry.Path), e);
+                        }
                     }
-                    catch (Exception e)
-                    {
-                        Smithbox.LogError(this, LOC.Get("MAP_Data_Failed_Write_BTPB", entry.Path), e);
-                    }
+                }
+                else
+                {
+                    Smithbox.LogError(this, LOC.Get("MAP_Data_Failed_Find_BTPB", entry.Path));
                 }
             }
         }
